Updated Naked checks for usage in Events.
Rather than calling the scene class, the Game_Interpreter has been
extended to support checks if the player, a member of the players party,
or an actor in the game is naked.

This means that you only need to add a script like isNaked? to see if the
player is naked when writing an event. No need to call an instantiated
object.

Game_Actor has been extended with the method naked? which returns true if
the body armor is not set (0).

Game_Actors has been extended with having to_a return @data (all actors
defined in the game). This is to have easy access to and avoid having to
rebuild an array of actors every time a query is made relating to the
actors in the database. The old method was to iterate over all possible
indicies and break when and index returned nil (except for index 0).

Extended Game_Interpreter with methods isNaked? and isAnyoneNaked?.
isNaked? can be passed with the index of the member you wish to check (0 =
leader) or the (case sensitive) name of the actor, e.g. isNaked? "Eva". If
nothing is provided, the party leader is the one checked by default.
isAnyoneNaked? checks if there is anyone in the party that is naked. This
is different from the previous implementation since it only checks the
actors that are actually in the party, rather than going through the actor
definitions in the database.

Usage could look like this (event):
> Conditional Branch: isNaked?("Eva")
  Text: Nice "outfit", Eva.
: Else
  > Conditional Branch: isAnyoneNaked?
    Text: Oh! I like her, Eva. Can I buy her?
  : Branch End
: Branch End

Function debug_put has been defined to recieve all error messages, and
will print them if the game is in DEBUG mode (i.e. running the game
through the editor)

All scripts has been added to "ATQ Naked" in the Script Editor.