Added ability to run Interpreter on any scene.
Added accessor for interpreter_queue for Scene_Base. By adding a
Game_Interpreter to the current scene ($scene), you can now do fun things
like adding calls to common events when the player isn't on the map.

Game_Actor has been extended to show this. Whenever the Game_Actor's
equipment is modified, Common Event 1 is called (if not testing). This
makes it possible both to provide information and keep the sprites
up-to-date without doing major changes to how the core game functions.

@event_id is set to the actor's id, to make it easer for to ID who was
changed. (e.g. @event_id = 1 -> equipment was changed for the first actor,
Eva.)

Game_Actor was modified to store @actor_id and to defined and add the
interpreter to the current $scene, whenever the equipment is modified.

Common Event 1 was updated to update the player's sprite whenever the
actors equipment is modified (through the equipment menu, or through other
means). Currently the cow, slave, normal dress and naked sprites are used,
and matched to the relevant armors.