Removed Editor check from data converter scripts.
Removed the checks to see if the editor is running before having the data
exported/imported by the script.
This is mostly to be able to use the exporter and commit frequently when
working with the game (i.e. make change in editor, run exporter, make
commit, continue working) without having to constantly close the editor as
required by start_rmvx.
Data importer is more tricky to work with since the editor loads maps into
memory when visiting them (not sure how it is with actors and that stuff).
At least you can try to import data now when the editor is running, but I
can't guarantee that the changes are reflected in the editor.