Added raise to Game_Interpreter and history to raise.
History allows for tracking when changes are done to attributes. This can be used for capturing repeatable (and exploitable) events, or if you want to have the impact of changes vary over time (e.g. each change can have a different impact on the attribute based on how long ago it happened).

Game_Troop
  setup_battle_event - will now pass the page for events played out

Game_Event
:page_id
  The current page of the event. This is made available for the Game_Interpreter to read, but will also be passed to the interpreter when the event autoruns.

Game_Interpreter
@page_id
  The current page number of the event being run
:raise(attribute, amount, max[, custom_context])
  Retrieves the context (map_id, event_id, page_id) + passed custom_context, and passes it on to the player's :raise method. Player is considered the leader of the game_party.
  Can be called using the script event command.

    raise("pervert", 1, 50) # add 1 to perversion, up to 50
    raise("pervert", 3, 60, "mySpecialContext") # add 3 to perversion, up to 60, for context "mySpecialContext"

Game_actor
:raise(attribute, amount, max[, context])
  context - value that uniquely identifies the attribute raise.
  If there's a history of the context used for raising the attribute before, it will lessen the change by the number of times called (2 times = delta / 2, 3 times = delta/3 and so on) until the it reaches 1, which will change the delta to 0.
:attribute_history
  ... is a hash containing changes through raise. Key is the context [map_id, event_id, page_id], and the value is when it happened in game time (in seconds).

Index (where on the page the script call is) is not used by default for
the context, since it may give you faulty result during development (where
the page position of the command may change). If it's known that the event's won't change, then having it would be better, because it wouldn't interfere in those cases where one page can do a number of changes to the same attribute (and you don't want to lose context). A workaround is to add a custom context by e.g. adding a digit.

    def wear_plugged_tail(do_wear) # script command on an event's page
      if do_wear
        raise("pervert", 5, 50, 1) # Add 5, up to 50, for context 1
      else
        raise("pervert", 3, 50, 2) # Add 5, up to 50, for context 2
      end
    end

    actor.pervert = 0
    wear_plugged_tail(false) # don't wear
      => 3 # pervert raised by 3 = 3 / 1
    wear_plugged_tail(true) # do wear
      => 8 # pervert raised by 5 = 5 / 1
    wear_plugged_tail(true) # do wear
      => 10 # pervert raised by 2 = 5 / 2

The extra context will only extend the existing contest (map, event, page)
so it can only be used to (un)link raise commands on a page.