Changed attribute_history internal storage.
attribute_history is now a hash with attribute name as key, and a array of changes as the value.
Each change is a hash with the keys :time (the game frame when it happened), :context (provided context, usually: map, event, page) and :value (the change itself positive value for raise, negative value for lowering).

A recommended improvement would be to index time and context in separate values, pointing at the relevant values. Another improvement would be to not have changes of 0 stored in the list.