Fixed some possible issues with achievements.
Added exclusive lock to save method, and made the method download the old definitions of cleared achievements, so it can be merged with the one stored in memory. This means that old unlocks of removed achievements remains and that achievements gets saved properly if multiple instances of the game is running.

  :load_save(if = nil)
    Can now be passed the file pointer. This is used when updating the content is up-to-date at save time, where the locked pointer is passed to the method.