Updated download links.
Changed so all links in the download are written in full (rather than being hidden in links).
Changed SVN address to use the new system of linking to the Git repository. The previous address is being deprecated by Github, and may cause issues during checkout. trunk was appended to the end of the address to avoid having new users download all of the branches, given how slow that would be.

`https://github.com/glqgamemaster/Gentle-Love-Quest.git` works just as well as `https://github.com/glqgamemaster/Gentle-Love-Quest` for downloading the project through SVN. I'm not sure if there could be possible conflicts in keeping the extension (`.git`) with some SVN clients. There probably isn't, but just to be on the safe side, I used the same format as the Github staff presented in their [post](https://github.com/blog/966-improved-subversion-client-support).